Update for better dependency handling
The compile classpaths are all resolving and the various
task dependeny graphs are correctly resolving. We should
be able to merge SpongeVanilla into SpongeCommon fairly
soon, but I'll hold off until we can define some better
understanding of how ModLauncher and coremods are needed
to allow us to perform some of the changes that we want
to be able to perform.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>